requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
django-cors-headers
djangorestframework-simplejwt
pandas
openpyxl
reportlab
Faker